Webscrapper (mostly) working now
It isn’t pulling the correct content for some variables, but that has
to do with the hosted emails I’m trying to scrape needing some IDs in
place. It still needs to be linked through the frontend. It still needs
to have a database created for it AND FOR THE LOVE OF GOD get it out of
the grunt file.